+++ changelog.md
@@ -1,6 +1,8 @@
 # Version 2025.1.2 (2025-01-06)
 
-- Identify channel of a system variable if name ends with channel address
+- Identify channel of a system variable:
+  - name ends with channel address
+  - name contains channel/device id
 
 # Version 2025.1.1 (2025-01-05)